Update.
1997-05-07 13:26  Andreas Schwab  <schwab@issan.informatik.uni-dortmund.de>

	* Makerules (lib): Put back dependency on lib-noranlib.
	($(subdirs-stamps)): Remove dependency on stamp.* again.

1998-05-07  Ulrich Drepper  <drepper@cygnus.com>

	* sysdeps/i386/fpu/bits/mathinline.h: Use long long int also for
	test in __pow2 implementation.

1998-05-07  Andreas Jaeger  <aj@arthur.rhein-neckar.de>

	* malloc/malloc.c (top_check): Add void as parameter to silent gcc.

+++ sysdeps/i386/fpu/bits/mathinline.h
@@ -231,7 +231,7 @@ __inline_mathcode (__sgn, __x, \
 __inline_mathcode (__pow2, __x, \
   register long double __value;						      \
   register long double __exponent;					      \
-  long int __p = (long int) __x;					      \
+  long long int __p = (long long int) __x;				      \
   if (__x == (long double) __p)						      \
     {									      \
       __asm __volatile__						      \
